fix: couponcampaign buy score

+++ frontend/plugins/loyalty_ui/src/modules/loyalties/settings/coupon/graphql/mutations/CouponMutations.tsx
@@ -15,7 +15,7 @@ export const CREATE_COUPON_CAMPAIGN = gql`
     $codeRule: JSON
     $restrictions: JSON
     $redemptionLimitPerUser: Int
-    $buyScore: Int
+    $buyScore: Float
     $charSet: [String]
   ) {
     couponCampaignAdd(
